add pom

+++ spring-boot-starter-saluki/src/main/java/com/quancheng/saluki/boot/runner/GrpcServiceRunner.java
@@ -56,8 +56,12 @@ public class GrpcServiceRunner implements DisposableBean, CommandLineRunner {
 
     @Override
     public void destroy() throws Exception {
-        rpcService.destroy();
-        applicationContext.destroy();
+        if (rpcService != null) {
+            rpcService.destroy();
+        }
+        if (applicationContext != null) {
+            applicationContext.destroy();
+        }
     }
 
     @Override
